+++ types/protocol-mapping.d.ts
@@ -486,6 +486,11 @@ export namespace ProtocolMapping {
          * Fired when a prerender attempt is completed.
          */
         'Page.prerenderAttemptCompleted': [Protocol.Page.PrerenderAttemptCompletedEvent];
+        /**
+         * TODO(crbug/1384419): Create a dedicated domain for preloading.
+         * Fired when a prefetch attempt is updated.
+         */
+        'Page.prefetchStatusUpdated': [Protocol.Page.PrefetchStatusUpdatedEvent];
         'Page.loadEventFired': [Protocol.Page.LoadEventFiredEvent];
         /**
          * Fired when same-document navigation happens, e.g. due to history API usage or anchor navigation.
@@ -704,6 +709,11 @@ export namespace ProtocolMapping {
          * list of player ids and all events again.
          */
         'Media.playersCreated': [Protocol.Media.PlayersCreatedEvent];
+        /**
+         * A device request opened a user prompt to select a device. Respond with the
+         * selectPrompt or cancelPrompt command.
+         */
+        'DeviceAccess.deviceRequestPrompted': [Protocol.DeviceAccess.DeviceRequestPromptedEvent];
     }
 
     export interface Commands {
@@ -4589,6 +4599,34 @@ export namespace ProtocolMapping {
          */
         'Media.disable': {
             paramsType: [];
+            returnType: void;
+        };
+        /**
+         * Enable events in this domain.
+         */
+        'DeviceAccess.enable': {
+            paramsType: [];
+            returnType: void;
+        };
+        /**
+         * Disable events in this domain.
+         */
+        'DeviceAccess.disable': {
+            paramsType: [];
+            returnType: void;
+        };
+        /**
+         * Select a device in response to a DeviceAccess.deviceRequestPrompted event.
+         */
+        'DeviceAccess.selectPrompt': {
+            paramsType: [Protocol.DeviceAccess.SelectPromptRequest];
+            returnType: void;
+        };
+        /**
+         * Cancel a prompt in response to a DeviceAccess.deviceRequestPrompted event.
+         */
+        'DeviceAccess.cancelPrompt': {
+            paramsType: [Protocol.DeviceAccess.CancelPromptRequest];
             returnType: void;
         };
     }