+++ types/protocol-proxy-api.d.ts
@@ -102,6 +102,8 @@ export namespace ProtocolProxyApi {
 
         Media: MediaApi;
 
+        DeviceAccess: DeviceAccessApi;
+
     }
 
 
@@ -3069,6 +3071,12 @@ export namespace ProtocolProxyApi {
          */
         on(event: 'prerenderAttemptCompleted', listener: (params: Protocol.Page.PrerenderAttemptCompletedEvent) => void): void;
 
+        /**
+         * TODO(crbug/1384419): Create a dedicated domain for preloading.
+         * Fired when a prefetch attempt is updated.
+         */
+        on(event: 'prefetchStatusUpdated', listener: (params: Protocol.Page.PrefetchStatusUpdatedEvent) => void): void;
+
         on(event: 'loadEventFired', listener: (params: Protocol.Page.LoadEventFiredEvent) => void): void;
 
         /**
@@ -3902,6 +3910,35 @@ export namespace ProtocolProxyApi {
          * list of player ids and all events again.
          */
         on(event: 'playersCreated', listener: (params: Protocol.Media.PlayersCreatedEvent) => void): void;
+
+    }
+
+    export interface DeviceAccessApi {
+        /**
+         * Enable events in this domain.
+         */
+        enable(): Promise<void>;
+
+        /**
+         * Disable events in this domain.
+         */
+        disable(): Promise<void>;
+
+        /**
+         * Select a device in response to a DeviceAccess.deviceRequestPrompted event.
+         */
+        selectPrompt(params: Protocol.DeviceAccess.SelectPromptRequest): Promise<void>;
+
+        /**
+         * Cancel a prompt in response to a DeviceAccess.deviceRequestPrompted event.
+         */
+        cancelPrompt(params: Protocol.DeviceAccess.CancelPromptRequest): Promise<void>;
+
+        /**
+         * A device request opened a user prompt to select a device. Respond with the
+         * selectPrompt or cancelPrompt command.
+         */
+        on(event: 'deviceRequestPrompted', listener: (params: Protocol.DeviceAccess.DeviceRequestPromptedEvent) => void): void;
 
     }
 }